Revert "Fix issues on ARM."  Causes issues on x86.
This reverts commit e0a603d33baa4e4940f0467dca70d06e8ced8457.

+++ Event.cpp
@@ -77,7 +77,7 @@ void Event::calculate(size_t numSims) {
 	std::mt19937_64 generator(rand_dev());
 	std::uniform_int_distribution<unsigned long> distribution(0x00000000, 0xffffffff);
 
-	unsigned long long randomPossibility = 0;
+	unsigned long randomPossibility = 0;
 
 	do {
 		randomPossibility = distribution(generator);
